LPS-135918 Fix problem updating Questions App Configuration

+++ modules/apps/questions/questions-web/src/main/java/com/liferay/questions/web/internal/configuration/persistence/listener/QuestionsConfigurationModelListener.java
@@ -198,13 +198,13 @@ public class QuestionsConfigurationModelListener
 
 		String name = "QUESTIONS_SERVICE_ACCESS_POLICY";
 
-		if (!enableAnonymousRead) {
-			SAPEntry sapEntry = _sapEntryService.getSAPEntry(
-				CompanyThreadLocal.getCompanyId(), name);
+		SAPEntry sapEntry = _sapEntryService.fetchSAPEntry(
+			CompanyThreadLocal.getCompanyId(), name);
 
+		if (!enableAnonymousRead && (_sapEntryService != null)) {
 			_sapEntryService.deleteSAPEntry(sapEntry);
 		}
-		else {
+		else if (enableAnonymousRead && (_sapEntryService == null)) {
 			String mbPackage = "com.liferay.message.boards.service.";
 
 			_sapEntryService.addSAPEntry(
